pyproject.toml
[project]
name = "atlassian-ai-assistant"
version = "2.0.0"
description = "AI-powered assistant that integrates with enterprise Jira and Confluence instances"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@company.com"}
]
keywords = ["jira", "confluence", "ai", "assistant", "atlassian", "mcp", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

# Core dependencies
dependencies = [
    # Claude SDK
    "anthropic>=0.40.0",

    # Claude Agent SDK (for MCP and Skills)
    "claude-agent-sdk>=0.1.0",

    # OpenAI Agents SDK (for MCP and Skills with OpenAI models)
    "openai-agents>=0.2.0",

    # MCP Server
    "mcp>=1.0.0",

    # Atlassian API Client
    "atlassian-python-api>=3.41.0",

    # HTTP Client
    "httpx>=0.27.0",

    # Environment Management
    "python-dotenv>=1.0.0",

    # JSON handling
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/zhuoli/confluence_assistant"
Documentation = "https://github.com/zhuoli/confluence_assistant/blob/main/README.md"
Repository = "https://github.com/zhuoli/confluence_assistant"
Issues = "https://github.com/zhuoli/confluence_assistant/issues"

[project.scripts]
atlassian-ai = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
